backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.28.0
tzlocal>=5.0
loguru==0.7.3
python-dateutil==2.9.0
pytz==2025.2